Add Pick cargo logic

+++ RobotBLL/Implementation/States/GameState.cs
@@ -8,7 +8,7 @@ namespace RobotBLL.Implementation.States
     public class GameState
     {
         public int CargoAmount { get; set; } = 0;
-        public float TotalCurrentPrice { get; set; } = 0;
+        public double TotalCurrentPrice { get; set; } = 0;
         public bool IsEnded { get; set; } = false;
         public Field GameField { get; set; }
